Upgrades gantry system + better structure
