+++ notes.md
@@ -1,0 +1,1 @@
+* The gantry system cannot load programs anymore after using it and break the connection
