+++ ros2gui/__init__.py
@@ -1,0 +1,1 @@
+# ros2gui package
